Move parse result mutations into their own files
