+++ lib/yarp.rb
@@ -248,193 +248,6 @@ module YARP
     def failure?
       !success?
     end
-
-    # CommentAttacher is a utility class to attach comments to locations in the AST
-    class CommentAttacher
-      attr_reader :parse_result
-
-      def initialize(parse_result)
-        @parse_result = parse_result
-      end
-
-      def attach!
-        parse_result.comments.each do |comment|
-          preceding, enclosing, following = nearest_targets(parse_result.value, comment)
-          target =
-            if comment.trailing?
-              preceding || following || enclosing || NodeTarget.new(parse_result.value)
-            else
-              # If a comment exists on its own line, prefer a leading comment.
-              following || preceding || enclosing || NodeTarget.new(parse_result.value)
-            end
-
-          target << comment
-        end
-      end
-
-      # A target for attaching comments that is based on a specific node
-      class NodeTarget
-        attr_reader :node
-
-        def initialize(node)
-          @node = node
-        end
-
-        def start_offset
-          node.location.start_offset
-        end
-
-        def end_offset
-          node.location.end_offset
-        end
-
-        def encloses?(comment)
-          start_offset <= comment.location.start_offset && comment.location.end_offset <= end_offset
-        end
-
-        def <<(comment)
-          node.location.comments << comment
-        end
-      end
-
-      # A target for attaching comments that is based on a location, which could be a part of a node. For example, the
-      # `end` token of a ClassNode
-      class LocationTarget
-        attr_reader :location
-
-        def initialize(location)
-          @location = location
-        end
-
-        def start_offset
-          location.start_offset
-        end
-
-        def end_offset
-          location.end_offset
-        end
-
-        def encloses?(comment)
-          false
-        end
-
-        def <<(comment)
-          location.comments << comment
-        end
-      end
-
-      private
-
-      # Responsible for finding the nearest targets to the given comment within the context of the given encapsulating
-      # node.
-      def nearest_targets(node, comment)
-        comment_start = comment.location.start_offset
-        comment_end = comment.location.end_offset
-
-        targets = []
-        node.deconstruct_keys(nil).each do |key, value|
-          next if key == :location
-
-          case value
-          when StatementsNode
-            targets.concat(value.body.map { |node| NodeTarget.new(node) })
-          when Node
-            targets << NodeTarget.new(value)
-          when Location
-            targets << LocationTarget.new(value)
-          when Array
-            targets.concat(value.map { |node| NodeTarget.new(node) }) if value.first.is_a?(Node)
-          end
-        end
-
-        targets.sort_by!(&:start_offset)
-        preceding = nil
-        following = nil
-
-        left = 0
-        right = targets.length
-
-        # This is a custom binary search that finds the nearest nodes to the given comment. When it finds a node that
-        # completely encapsulates the comment, it recursed downward into the tree.
-        while left < right
-          middle = (left + right) / 2
-          target = targets[middle]
-
-          target_start = target.start_offset
-          target_end = target.end_offset
-
-          if target.encloses?(comment)
-            # The comment is completely contained by this target. Abandon the binary search at this level.
-            return nearest_targets(target.node, comment)
-          end
-
-          if target_end <= comment_start
-            # This target falls completely before the comment. Because we will never consider this target or any targets
-            # before it again, this target must be the closest preceding target we have encountered so far.
-            preceding = target
-            left = middle + 1
-            next
-          end
-
-          if comment_end <= target_start
-            # This target falls completely after the comment. Because we will never consider this target or any targets
-            # after it again, this target must be the closest following target we have encountered so far.
-            following = target
-            right = middle
-            next
-          end
-
-          # This should only happen if there is a bug in this parser.
-          raise "Comment location overlaps with target location"
-        end
-
-        [preceding, NodeTarget.new(node), following]
-      end
-    end
-
-    # Attach the list of comments to their respective locations in the AST
-    def attach_comments!
-      CommentAttacher.new(self).attach!
-    end
-
-    # Keep in sync with Java MarkNewlinesVisitor
-    class MarkNewlinesVisitor < YARP::Visitor
-      def initialize(newline_marked)
-        @newline_marked = newline_marked
-      end
-
-      def visit_block_node(node)
-        old_newline_marked = @newline_marked
-        @newline_marked = Array.new(old_newline_marked.size, false)
-        begin
-          super(node)
-        ensure
-          @newline_marked = old_newline_marked
-        end
-      end
-      alias_method :visit_lambda_node, :visit_block_node
-
-      def visit_if_node(node)
-        node.set_newline_flag(@newline_marked)
-        super(node)
-      end
-      alias_method :visit_unless_node, :visit_if_node
-
-      def visit_statements_node(node)
-        node.body.each do |child|
-          child.set_newline_flag(@newline_marked)
-        end
-        super(node)
-      end
-    end
-    private_constant :MarkNewlinesVisitor
-
-    def mark_newlines
-      newline_marked = Array.new(1 + @source.offsets.size, false)
-      visitor = MarkNewlinesVisitor.new(newline_marked)
-      value.accept(visitor)
-      value
-    end
   end
 
   # This represents a token from the Ruby source.
@@ -777,6 +590,9 @@ require_relative "yarp/ripper_compat"
 require_relative "yarp/serialize"
 require_relative "yarp/pack"
 require_relative "yarp/pattern"
+
+require_relative "yarp/parse_result/comments"
+require_relative "yarp/parse_result/newlines"
 
 if RUBY_ENGINE == "ruby" and !ENV["YARP_FFI_BACKEND"]
   require "yarp/yarp"
